examples: separate example for bmfont from 'text' to a new one 'font'
