+++ examples/text/src/main.cpp
@@ -14,7 +14,6 @@
 #include "bn_sprite_text_generator.h"
 
 #include "fixed_32x64_sprite_font.h"
-#include "user_variable_32x32_sprite_font.h"
 
 #include "bn_sprite_items_variable_8x16_font_red.h"
 #include "bn_sprite_items_variable_8x16_font_blue.h"
@@ -27,7 +26,7 @@
 
 namespace
 {
-    constexpr bn::fixed text_y_inc = 12;
+    constexpr bn::fixed text_y_inc = 14;
     constexpr bn::fixed text_y_limit = (bn::display::height() / 2) - text_y_inc;
 
     void text_scene()
@@ -216,30 +215,6 @@ namespace
             bn::core::update();
         }
     }
-
-    void user_font_text_scene()
-    {
-        bn::sprite_text_generator text_generator(common::variable_8x16_sprite_font);
-        text_generator.set_center_alignment();
-
-        bn::vector<bn::sprite_ptr, 32> text_sprites;
-        text_generator.generate(0, -text_y_limit, "BMFont text (user)", text_sprites);
-        text_generator.generate(0, text_y_limit, "START: go to next scene", text_sprites);
-
-        bn::sprite_text_generator user_font_text_generator(user_variable_32x32_sprite_font);
-        user_font_text_generator.set_center_alignment();
-
-        bn::vector<bn::sprite_ptr, 32> user_font_text_sprites;
-        user_font_text_generator.generate(0, -48, "这是用户自定义", user_font_text_sprites);
-        user_font_text_generator.generate(0, -16, "字体。", user_font_text_sprites);
-        user_font_text_generator.generate(0, 16, "ユーザー設定の", user_font_text_sprites);
-        user_font_text_generator.generate(0, 48, "フォントです。", user_font_text_sprites);
-
-        while(! bn::keypad::start_pressed())
-        {
-            bn::core::update();
-        }
-    }
 }
 
 int main()
@@ -269,9 +244,6 @@ int main()
         bn::core::update();
 
         utf8_text_scene();
-        bn::core::update();
-
-        user_font_text_scene();
         bn::core::update();
     }
 }
